Create SubjectTest.jsx
This is the SubjectTest module for performing tests. Is starts with the start button then a notes tab, when we click on the continue button it shows the quiz box and there the quiz takes place, only when the user complete the quiz or the time ends it shows the result box.